refactor: extract Deck Item from Deck page #103
