+++ src/pages/Deck/subcomponents/DeckList/DeckItem/index.ts
@@ -1,0 +1,1 @@
+export * from './DeckItem';
