+++ src/pages/Deck/subcomponents/DeckList/DeckList.style.tsx
@@ -4,17 +4,6 @@ export const DeckListContainer = styled.ul`
   min-height: 345rem;
 `;
 
-export const DeckItemContainer = styled.li`
-  min-height: 23.25rem;
-`;
-
-export const CardContainer = styled.div`
-  display: flex;
-  gap: 1.5rem;
-  flex-wrap: wrap;
-  flex-grow: 1;
-`;
-
 export const LoaderContainer = styled.div`
   display: flex;
   align-items: center;
